+++ Client/src/i18n/en-us/errors.js
@@ -1,0 +1,3 @@
+export default {
+  CaptchaValidationError: "Ошибка проверки капчи"
+}
