+++ Client/src/modules/account/ResetPasswordSetNew.vue
@@ -100,12 +100,7 @@
         }).then( () => {
           this.done = true;
         }).catch(error => {
-          this.$q.notify({
-            message: error.response.data.errorText,
-            timeout: 5000,
-            color: 'negative',
-            position: 'top'
-          });
+          this.$errorNotify(error.response.data);
           this.submitting = false;
         });
       }
