+++ SunEngine.Admin/Controllers/CategoriesAdminController.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Mvc;
 using SunEngine.Admin.Managers;
 using SunEngine.Admin.Presenters;
 using SunEngine.Commons.Cache;
+using SunEngine.Commons.Cache.Services;
 using SunEngine.Commons.Models;
 
 namespace SunEngine.Admin.Controllers
