+++ SunEngine.Admin/Managers/CategoriesAdminManager.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using System.Transactions;
 using LinqToDB;
 using SunEngine.Commons.Cache;
+using SunEngine.Commons.Cache.Services;
 using SunEngine.Commons.DataBase;
 using SunEngine.Commons.Models;
 using SunEngine.Commons.Services;
