+++ SunEngine.Commons/Configuration/AddServices/AddStores.cs
@@ -1,5 +1,6 @@
 using Microsoft.Extensions.DependencyInjection;
 using SunEngine.Commons.Cache;
+using SunEngine.Commons.Cache.Services;
 using SunEngine.Commons.DataBase;
 
 namespace SunEngine.Commons.Configuration.AddServices
