+++ SunEngine.Commons/Controllers/BaseController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
 using SunEngine.Commons.Cache;
 using SunEngine.Commons.Cache.CacheModels;
+using SunEngine.Commons.Cache.Services;
 using SunEngine.Commons.Managers;
 using SunEngine.Commons.Models;
 using SunEngine.Commons.Security;
