+++ SunEngine.Commons/Presenters/CategoriesPresenter.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using SunEngine.Commons.Cache;
 using SunEngine.Commons.Cache.CacheModels;
+using SunEngine.Commons.Cache.Services;
 using SunEngine.Commons.Security;
 
 namespace SunEngine.Commons.Presenters
