+++ SunEngine.Commons/Security/JwtBlackListService.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using LinqToDB;
 using Microsoft.Extensions.Options;
 using SunEngine.Commons.Cache;
+using SunEngine.Commons.Cache.Services;
 using SunEngine.Commons.Configuration.Options;
 using SunEngine.Commons.DataBase;
 using SunEngine.Commons.Models.Authorization;
