+++ SunEngine.Commons/Security/SunClaimsPrincipal.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Security.Claims;
 using SunEngine.Commons.Cache;
 using SunEngine.Commons.Cache.CacheModels;
+using SunEngine.Commons.Cache.Services;
 
 namespace SunEngine.Commons.Security
 {
